model_deployment/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
scikeras
tensorflow
numpy
google-cloud-aiplatform